Remove core vscode.lm getConsoleContent tool
The core getConsoleContent tool was exposed to chat clients via
vscode.lm. Now that Posit Assistant reads console history natively
through the positron.ai.getConsoleContent extension API, this tool is
vestigial and contradicts the direction of moving off vscode.lm tools,
so remove it (tool data, class, registration, and tool-name enum entry).

The projection logic (projectExecutionEntriesToConsoleContent) stays in
the execution history service, still used by the extension API, and gains
a direct unit test in place of the removed tool's tests.
