+++ src/vs/workbench/api/test/common/positron/positronToolFilter.vitest.ts
@@ -76,14 +76,14 @@ describe('getEnabledTools', () => {
 		expect(getEnabledTools(inEditor, tools, true)).toEqual([]);
 	});
 
-	it('keeps read-only getConsoleContent available in Ask mode while excluding executeCode', () => {
+	it('keeps a read-only session tool available in Ask mode while excluding executeCode', () => {
 		const tools = [
 			tool('executeCode', ['positron-assistant', 'requires-session']),
-			tool('getConsoleContent', ['positron-assistant', 'requires-session']),
+			tool('inspectVariables', ['positron-assistant', 'requires-session']),
 		];
 		// Ask mode is the Chat participant (not agent mode); executeCode is
-		// gated to agent mode, but the read-only console tool stays enabled.
-		expect(getEnabledTools(request(), tools, true, ParticipantID.Chat)).toEqual(['getConsoleContent']);
+		// gated to agent mode, but a read-only session tool stays enabled.
+		expect(getEnabledTools(request(), tools, true, ParticipantID.Chat)).toEqual(['inspectVariables']);
 	});
 
 	it('disables all tools for the terminal participant', () => {
