+++ src/vs/workbench/contrib/positronAssistant/common/positronAssistantToolNames.ts
@@ -12,7 +12,6 @@
  */
 export enum PositronAssistantToolName {
 	ExecuteCode = 'executeCode',
-	GetConsoleContent = 'getConsoleContent',
 	GetTableSummary = 'getTableSummary',
 	GetPlot = 'getPlot',
 	InspectVariables = 'inspectVariables',
